First draft.
Integrated the original HTML into the Jekyll layouts/includes style.

Only downside is an ugly if/else to ensure the <body> gets the correct
style in the homepage as opposed to a post page; the effect is to place
the footer in the sidebar on the homepage but center it on post pages
(where there is no sidebar).